--- a/.env.example
+++ b/.env.example
@@ -99,3 +99,5 @@ NEXT_PUBLIC_CHAOS_INTENSITY=0.3
 NEXT_PUBLIC_CHAOS_TARGETS=brand_hero_cta
 # Enable console logs for Chaos Mode
 NEXT_PUBLIC_CHAOS_DEBUG=false
+AI_GATEWAY_API_KEY=vck_1PDTpNm71DxHqMHp2U9zC8HeudDf1RCyQnS5vm1DWeLMqPHSbJ27nVyn
+AI_GATEWAY_URL=https://gateway.ai.vercel.com/api/v1