--- a/app/api/embeddings/route.ts
+++ b/app/api/embeddings/route.ts
@@ -6,7 +6,7 @@ export const runtime = 'nodejs'
 
 const openai = new OpenAI({
   apiKey: process.env.AI_GATEWAY_API_KEY || process.env.OPENAI_API_KEY,
-  baseURL: process.env.AI_GATEWAY_API_KEY ? (process.env.AI_GATEWAY_URL || 'https://gateway.ai.vercel.com/api/v1') : undefined,
+  baseURL: process.env.AI_GATEWAY_API_KEY ? process.env.AI_GATEWAY_URL : undefined,
 })
 
 export async function POST(request: NextRequest) {