--- a/app/api/rag/query/route.ts
+++ b/app/api/rag/query/route.ts
@@ -33,9 +33,17 @@ const INDEX_PATH = path.join(DATA_DIR, 'index.json');
 // Helper functions
 function buildAnswerPrompt(question: string, contexts: { text: string; metadata: Record<string, any> }[]) {
   const contextStr = contexts.map((c) => `Source: ${c.metadata?.source_name || 'Unknown'}\n${c.text}`).join('\n\n---\n\n');
-  const system = `You are an expert design leader and product strategist. Answer based on the provided context. If the context doesn't contain relevant information, say so clearly.
+  const system = `You are answering as a senior design leader and product strategist in a job interview context. 
 
-Context:
+CRITICAL INSTRUCTIONS:
+- Always respond in FIRST PERSON ("I have...", "My experience includes...", "I led...")
+- Never use third person ("You have...", "Your experience...")
+- Draw directly from the provided context about your background and projects
+- Be specific about projects, metrics, and outcomes from your experience
+- If the context doesn't contain relevant information, acknowledge what you can speak to from your background
+- Maintain a confident, professional tone appropriate for a senior leadership role
+
+Context about your background:
 ${contextStr}
 
 Question: ${question}`;