--- a/app/components/grid/useGridState.ts
+++ b/app/components/grid/useGridState.ts
@@ -1,0 +1,7 @@
+'use client'
+
+import { useGridState } from './GridStateProvider'
+
+export function useGridState() {
+  return useGridState()
+}