@hretheum 2025-09-29
## ✅ Build Fix + T-Mobile Added

**Changes:**
- Added `dotenv` to dependencies (fixes build error)
- Added T-Mobile job posting with personalized suggestions

**T-Mobile Suggestions:**
1. Tell me about your experience using **Miro** in design projects.
2. How did you manage your team at **ING Business** effectively?
3. What does **autonomy** in your work mean to you?
4. Can you describe your process for creating a **product roadmap**?
5. How do you stay updated on **FinTech trends and regulations**?

**Total brands with personalized suggestions: 8**
- Blix, Code & Pepper, Depot, EFI, Fluxon, Intent, SOFTSWISS, T-Mobile

Build should now pass! ✅