@hretheum 2025-09-29
## ⚡ Performance Optimization Added

**New commits:**
- Lazy loading for below-the-fold components
- Bundle size optimization (SWC minify, package imports)
- Image optimization config (AVIF/WebP)
- Resource hints (preconnect, dns-prefetch)

**Expected improvements:**
- LCP: 7069ms → ~3500ms (50% improvement)
- TTI: 6248ms → ~4000ms (35% improvement)
- TBT: 795ms → ~350ms (55% improvement)

**Changes:**
- `app/page.tsx` - Dynamic imports for 7 components
- `next.config.js` - Performance config
- `app/layout.tsx` - Resource hints

Tested on dev - all working! ✅